add subclass hook to CollectionMeta to fix tests
Weirest problem I've seen in awhile. Tests were failing because Jinja
apparently does a subclass comparison with all mapping types declared
anywhere in imported code. Because I failed to implement the subclass
hook on the CollectionMeta class, the issubclass call was failing.
However, the traceback was far from obvious and determining the error
was rather difficult.