remote test dir re #51
Nothing in the test dir is currently in use, so I am removing it. The
files in the payloads dir _may_ be useful in the future, but if so we
can always recover them from the git history. I figure it is better to
clean them up now than to keep them and end up not using them at all.