rework build test to not have copies of all files
When testing the files generated by the build step, we really only care
to diff the content of the serverless.yml file. For all other files we
can simply maintain a list of the shasums and sizes, and any differences
will tell us if a file has been added, removed, or changed.