cli dedup aliased and unaliased matches
When using short command matching, such as `c` for `cf` or `cloudformation`,
if a command and its alias are both matched we need to deduplicate the
matches after resolving aliases. If we do not do this, we will get an
error, for example `c` will error despite `cf` and `cloudformation`
being the same command.