throw error instead of returning None from EventDB _query
Reverts the changes to the API functions to handle null responses from
eventdb queries (which were added to appease type checking), with some
further simplifications. The _query method on EventDB will now raise an
exception if called when events are disabled. This is explicitly handled
in the API and ignored.