Add lambda request id to pre/post batch logs (#361)
* feat: add lambda request id to pre/post batch logs

* chore: update CHANGELOG

* docs: Add section on Lambda logs and filtering by request ID

* fix: update aiohttp in uv.lock

* fix: disable building wheels when building lambda dist

This prevents the problem of building a wheel on
one achitecture (x86_64 on the github runners) and
then using that wheel to install on another
architecture (e.g, aarch64).

Adds --no-build flag which requires pre-built wheels
to exist, failing fast if they're not available.

* fix: build a pure python wheel for cirrus-geo First

Allows us to install with --no-build

* revert: remove the --no-build flag from uv sync command

We can rely on uv sync to fail if it builds a wheel
(because one is not available) and that built wheel
is not compatible with the architecture of the target
platform.

* refactor: do not instantiate loggers inside lambda handlers

- Uses a reset_extra method for use inside handlers instead
- Allows passing additional "extra" on individual log calls

* chore: update CHANGELOG

* refactor: move get_task_logger logic to the __init__ of CirrusLoggerAdapter

- Also adds types to the class method signatures

* chore: update CHANGELOG